refactor directorios
