+++ app/src/main/java/com/programabit/mediguard/data/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.data;
 
 public class LoginRequest {
