+++ app/src/main/java/com/programabit/mediguard/domain/GuardsViewModel.java
@@ -1,10 +1,12 @@
-package com.programabit.mediguard.rest;
+package com.programabit.mediguard.domain;
 
 import android.app.Application;
 import android.util.Log;
 
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.MutableLiveData;
+
+import com.programabit.mediguard.data.GuardRestRepository;
 
 import java.util.List;
 import java.util.concurrent.ExecutionException;
