+++ app/src/main/java/com/programabit/mediguard/ui/AvaibleGuardsListAdapter.java
@@ -1,13 +1,11 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
 
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.ListAdapter;
-import com.programabit.mediguard.rest.GuardDto;
-
-import java.util.List;
+import com.programabit.mediguard.domain.GuardDto;
 
 public class AvaibleGuardsListAdapter extends ListAdapter<GuardDto, AvaibleGuardsViewHolder> {
