+++ app/src/main/java/com/programabit/mediguard/ui/AvaibleGuardsViewHolder.java
@@ -1,10 +1,12 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.programabit.mediguard.R;
 
 public class AvaibleGuardsViewHolder extends RecyclerView.ViewHolder {
