+++ app/src/main/java/com/programabit/mediguard/ui/ContactActivity.java
@@ -1,24 +1,25 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
+import com.programabit.mediguard.R;
 
-public class UserPhotoActivity extends AppCompatActivity {
+public class ContactActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_user_photo);
+        setContentView(R.layout.activity_contact);
 
-        // Toolbar
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        appToolbar(toolbar, R.string.acivity_name_update_photo,true);
+        androidx.appcompat.widget.Toolbar toolbar = findViewById(R.id.toolbar);
+        appToolbar(toolbar, R.string.activity_name_contact,true);
     }
 
     // AppBar (toolbar y menu):
@@ -33,8 +34,6 @@ public class UserPhotoActivity extends AppCompatActivity {
         int itemId = item.getItemId();
         if (itemId == R.id.mSettings) {
             startActivity(new Intent(this, UserSettingsActivity.class));
-        } else if (itemId == R.id.mContact) {
-            startActivity(new Intent(this, ContactActivity.class));
         } else if (itemId == R.id.mAbout) {
             startActivity(new Intent(this, AboutActivity.class));
         }
@@ -48,5 +47,11 @@ public class UserPhotoActivity extends AppCompatActivity {
             getSupportActionBar().setSubtitle(activity_name);
         }
         getSupportActionBar().setDisplayHomeAsUpEnabled(enable);
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        onBackPressed();
+        return true;
     }
 }
