+++ app/src/main/java/com/programabit/mediguard/ui/GuardsFactory.java
@@ -1,4 +1,4 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
 
 import android.app.Application;
 
@@ -6,7 +6,7 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.programabit.mediguard.rest.GuardsViewModel;
+import com.programabit.mediguard.domain.GuardsViewModel;
 
 import java.util.concurrent.ExecutionException;
