+++ app/src/main/java/com/programabit/mediguard/ui/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
 
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
@@ -14,7 +14,10 @@ import android.widget.Button;
 import android.widget.Toast;
 
 import com.google.android.material.textfield.TextInputEditText;
-import com.programabit.mediguard.rest.ApiClient;
+import com.programabit.mediguard.R;
+import com.programabit.mediguard.data.ApiClient;
+import com.programabit.mediguard.data.LoginRequest;
+import com.programabit.mediguard.data.LoginResponse;
 
 import androidx.appcompat.app.AppCompatActivity;
