+++ app/src/main/java/com/programabit/mediguard/ui/MainActivity.java
@@ -1,8 +1,10 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.programabit.mediguard.R;
 
 public class MainActivity extends AppCompatActivity {
