+++ app/src/main/java/com/programabit/mediguard/ui/MyGuardsActivity.java
@@ -1,4 +1,4 @@
-package com.programabit.mediguard;
+package com.programabit.mediguard.ui;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
@@ -16,8 +16,9 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.programabit.mediguard.rest.GuardDto;
-import com.programabit.mediguard.rest.GuardsViewModel;
+import com.programabit.mediguard.R;
+import com.programabit.mediguard.domain.GuardDto;
+import com.programabit.mediguard.domain.GuardsViewModel;
 
 import java.util.concurrent.ExecutionException;
