+++ app/src/main/java/com/programabit/mediguard/ui/UserPhotoActivity.java
@@ -1,23 +1,26 @@
-package com.programabit.mediguard;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
+package com.programabit.mediguard.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
-public class ContactActivity extends AppCompatActivity {
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+import com.programabit.mediguard.R;
+
+public class UserPhotoActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_contact);
+        setContentView(R.layout.activity_user_photo);
 
-        androidx.appcompat.widget.Toolbar toolbar = findViewById(R.id.toolbar);
-        appToolbar(toolbar, R.string.activity_name_contact,true);
+        // Toolbar
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        appToolbar(toolbar, R.string.acivity_name_update_photo,true);
     }
 
     // AppBar (toolbar y menu):
@@ -32,6 +35,8 @@ public class ContactActivity extends AppCompatActivity {
         int itemId = item.getItemId();
         if (itemId == R.id.mSettings) {
             startActivity(new Intent(this, UserSettingsActivity.class));
+        } else if (itemId == R.id.mContact) {
+            startActivity(new Intent(this, ContactActivity.class));
         } else if (itemId == R.id.mAbout) {
             startActivity(new Intent(this, AboutActivity.class));
         }
@@ -45,11 +50,5 @@ public class ContactActivity extends AppCompatActivity {
             getSupportActionBar().setSubtitle(activity_name);
         }
         getSupportActionBar().setDisplayHomeAsUpEnabled(enable);
-    }
-
-    @Override
-    public boolean onSupportNavigateUp() {
-        onBackPressed();
-        return true;
     }
 }
