+++ app/src/main/res/layout/activity_login.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical"
-    tools:context=".LoginActivity"
+    tools:context=".ui.LoginActivity"
     android:layout_width="fill_parent"
     android:layout_height="match_parent"
     android:layout_marginTop="0dp"
