+++ app/src/main/res/layout/activity_my_guards.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_height="match_parent"
     android:layout_width="wrap_content"
-    tools:context=".MyGuardsActivity"
+    tools:context=".ui.MyGuardsActivity"
     android:background="@color/colorBlanco">
 
     <ScrollView
