+++ app/src/main/res/layout/activity_user_settings.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UserSettingsActivity">
+    tools:context=".ui.UserSettingsActivity">
 
     <ScrollView
         android:id="@+id/scrollView1"
